Dev: Add ExcelFileInserter and adjust the logic in FileInserter, clean the code.
